--- src/Lexer-02.js
@@ -1,19 +1,0 @@
-// const tokens = [
-//   {type:"string", value: "la31"},
-//   {type:"string", value: "la32"},
-//   {type:"string", value: "la33"},
-//   {type:"null", value: null},
-//   {type: "undefined", value: undefined},
-//   {type:"number", value: 12},
-// ]
-
-const tokens = [
-  {type: "{", value: "{"},
-  {type: "prop-key", value: "tokens"},
-  {type: "object-spearator", value: ":"},
-  {type: "value", value: },
-  {type: "}", value: "}"},
-  {type: "[", value: "["},
-  {type: "]", value: "]"},
-  {type: ",", value: ","},
-]
