--- src/Lexer.js
@@ -1,17 +1,0 @@
-// null, undefined, true, false, 
-// ""(string), "123"(string), 123(Number), 
-
-
-const isString = () => {}
-const isNumber = () => {}
-const isNull = () => {}
-const isUndefined = () => {}
-const isBoolean = () => {}
-const isTokenSeparator = () => {}
-
-// {}
-// const isObject
-// (key) : (value)
-
-// []
-// const isArray
